refactor: Merge utils files
