+++ dpypelines/pipeline/shared/error_handler_module.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 from dpytools.logging.logger import DpLogger
 
-from dpypelines.pipeline.shared.utils import get_email_client
+from dpypelines.pipeline.utils import get_email_client
 from dpypelines.pipeline.utils import get_notifier
 
 logger = DpLogger("data-ingress-pipelines")
