+++ dpypelines/pipeline/shared/notification.py
@@ -3,7 +3,7 @@ from abc import ABC, abstractmethod
 
 from dpytools.slack.slack import SlackMessenger
 
-from dpypelines.pipeline.shared.utils import (
+from dpypelines.pipeline.utils import (
     get_commit_id,
     get_environment,
     get_local_time,
