+++ tests/pipelines/pipeline/shared/test_shared_utils.py
@@ -2,7 +2,7 @@ import subprocess
 
 import pytest
 
-from dpypelines.pipeline.shared.utils import (
+from dpypelines.pipeline.utils import (
     get_commit_id,
     get_mimetype,
     get_submitter_email,
